+++ InputGlyphs/Assets/InputGlyphs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.2.2] - Development
+### Modified
+- Supported `DISABLESTEAMWORKS` Assembly Definition.
+- Changed to use the parent path if the image cannot be found.
+  - Example: `leftStick/x` -> `leftStick`
+
 ## [1.2.1] - 2024-10-31
 ### Modified
 - Modified to automatically collect PlayerInput when it is not set.